jolt-claims: own lattice packing semantics
Move Jolt lattice family IDs, layout derivation, and validity formulas into jolt-claims so verifier code only adapts them to physical packing IDs.
