+++ crates/jolt-claims/src/protocols/jolt/mod.rs
@@ -3,6 +3,44 @@ pub mod formulas;
 mod ids;
 mod relation;
 
+pub use formulas::lattice::JoltPackingFamilyId;
+pub use formulas::lattice::{
+    advice_bytes_validity_requirement, bytecode_imm_canonical_bytes_requirement,
+    bytecode_validity_requirements, packing_validity_digest, program_image_validity_requirement,
+    unsigned_inc_validity_requirements,
+};
+pub use formulas::lattice::{
+    byte_decode_terms, bytecode_chunk_lattice_view_formula,
+    bytecode_rd_present_lattice_view_formula, bytecode_store_flag_lattice_view_formula,
+    little_endian_byte_decode_terms, symbol_decode_terms,
+    unsigned_inc_lower_value_lattice_view_formula, unsigned_inc_lower_value_terms,
+    unsigned_inc_msb_lattice_view_formula, weighted_byte_decode_terms, weighted_symbol_terms,
+};
+pub use formulas::lattice::{
+    derive_jolt_lattice_packed_validity_requirements, derive_jolt_lattice_packed_witness_layout,
+    jolt_advice_kind, lattice_validity_requirements_for_packed_witness_layout, layout_has_advice,
+    layout_has_field_rd_inc, packed_alphabet_with_size, packed_family_is_precommitted,
+    packing_advice_kind,
+};
+pub use formulas::lattice::{
+    final_opening_lattice_requirement, inc_virtualization_inc_opening,
+    inc_virtualization_input_openings, inc_virtualization_output_openings,
+    inc_virtualization_relation, inc_virtualization_store_opening, unsigned_inc_chunk_opening,
+    unsigned_inc_chunk_reconstruction_relation, unsigned_inc_claim_reduction_relation,
+    unsigned_inc_input_opening, unsigned_inc_lower_chunk_count, unsigned_inc_msb_opening,
+    unsigned_inc_opening,
+};
+pub use formulas::lattice::{
+    inc_virtualization_claim, unsigned_inc_chunk_reconstruction_claim,
+    unsigned_inc_claim_reduction_claim, unsigned_inc_msb_booleanity_claim,
+};
+pub use formulas::lattice::{
+    FieldRdIncPacking, JoltLatticeLayoutError, JoltLatticePackingInputs, JoltLatticeValidityInputs,
+};
+pub use formulas::lattice::{
+    LatticeFinalOpeningRequirement, PackingValidityDigest, PackingValidityKind,
+    PackingValidityRequirement, PackingViewFormula, PackingViewTerm, UNSIGNED_INC_BITS,
+};
 pub use formulas::{
     claim_reductions::advice::AdviceClaimReductionLayout,
     claim_reductions::bytecode::BytecodeClaimReductionLayout,
@@ -17,25 +55,6 @@ pub use formulas::{
         TraceDimensions, TracePolynomialOrder,
     },
     error::{JoltFormulaDimensionsError, JoltFormulaPointError},
-    lattice::{
-        advice_bytes_validity_requirement, byte_decode_terms, bytecode_chunk_lattice_view_formula,
-        bytecode_imm_canonical_bytes_requirement, bytecode_rd_present_lattice_view_formula,
-        bytecode_store_flag_lattice_view_formula, bytecode_validity_requirements,
-        final_opening_lattice_requirement, inc_virtualization_claim,
-        inc_virtualization_inc_opening, inc_virtualization_input_openings,
-        inc_virtualization_output_openings, inc_virtualization_relation,
-        inc_virtualization_store_opening, little_endian_byte_decode_terms, packing_validity_digest,
-        program_image_validity_requirement, symbol_decode_terms, unsigned_inc_chunk_opening,
-        unsigned_inc_chunk_reconstruction_claim, unsigned_inc_chunk_reconstruction_relation,
-        unsigned_inc_claim_reduction_claim, unsigned_inc_claim_reduction_relation,
-        unsigned_inc_input_opening, unsigned_inc_lower_chunk_count,
-        unsigned_inc_lower_value_lattice_view_formula, unsigned_inc_lower_value_terms,
-        unsigned_inc_msb_booleanity_claim, unsigned_inc_msb_lattice_view_formula,
-        unsigned_inc_msb_opening, unsigned_inc_opening, unsigned_inc_validity_requirements,
-        weighted_byte_decode_terms, weighted_symbol_terms, LatticeFinalOpeningRequirement,
-        PackingFamilyId, PackingValidityDigest, PackingValidityKind, PackingValidityRequirement,
-        PackingViewFormula, PackingViewTerm, UNSIGNED_INC_BITS,
-    },
 };
 pub use ids::{
     AdviceClaimReductionPublic, BooleanityChallenge, BooleanityPublic,
